Added new view controller to view todos in detail.
Added delegate protocol and function to move from the home view controller, to the todo detail view controller, however I've had to pass in the home view controller to its data source as a property as a hack to get this to work. Look for a better way to achieve this later.